+++ src/main/java/com/picotrake/API/repository/SuscripcionRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.picotrake.API.model.Suscripciones;
 
-public interface SuscripcionRepository extends JpaRepository<Suscripciones, Integer> {
-    Optional<Suscripciones> findById(int id);
+public interface SuscripcionRepository extends JpaRepository<Suscripciones, Long> {
+    Optional<Suscripciones> findById(Long id);
 }
